Rename ExamplePlugin to DataAnalysisToolkit.
