+++ wrapping/python/plugins/DataAnalysisToolkit/CalculateHistogramFilter.py
